+++ common/src/main/java/com/github/teamfossilsarcheology/fossil/inventory/CultureVatMenu.java
@@ -21,6 +21,7 @@ public class CultureVatMenu extends AbstractContainerMenu {
     public static final int FUEL_SLOT_ID = 1;
     public static final int OUTPUT_SLOT_ID = 2;
     public static final int CULTIVATION_DURATION = Version.debugEnabled() ? 100 : 6000;
+    public static final int DEFAULT_FUEL_DURATION = Version.debugEnabled() ? 1000 : 6000;
 
     private final Container container;
     private final ContainerData containerData;