+++ common/src/main/java/com/github/teamfossilsarcheology/fossil/recipe/MultiOutputAndSlotsRecipe.java
@@ -93,7 +93,7 @@ public abstract class MultiOutputAndSlotsRecipe implements Recipe<Container> {
         return id;
     }
 
-    public static class Serializer<T extends MultiOutputAndSlotsRecipe> extends AbstractRecipeSerializer<T> {
+    public static abstract class Serializer<T extends MultiOutputAndSlotsRecipe> extends AbstractRecipeSerializer<T> {
         protected final Constructor<T> constructor;
 
         protected Serializer(Constructor<T> constructor) {